Separate common.py and fold validate_raw_bucket_structure into scripts assisted with Claude
