+++ util/data_promotion/clean_wdl_raw_buckets
@@ -18,13 +18,10 @@ import sys
 import re
 from collections import defaultdict
 
-# --- resolve shared/ modules from new subfolder location ---
 import os, sys
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "shared"))
-from common import (
-    gremove,
-    unembargoed_dev_buckets_and_workflow_version_outputs,
-)
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "common"))
+from gcloud_ops import gremove
+from release_ops import unembargoed_dev_buckets_and_workflow_version_outputs
 
 
 FILTERED_TERMS = ["gcloud storage rm", "Removing"]
