+++ util/data_promotion/promote_raw_data
@@ -6,21 +6,23 @@ import re
 import logging
 import subprocess
 from google.cloud import storage
-# --- resolve shared/ modules from new subfolder location ---
+
 import os, sys
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "shared"))
-from common import (
-	releases_df,
-	completed_platforming_raw_buckets,
-	unembargoed_dev_buckets_and_workflow_version_outputs,
-	embargoed_dev_buckets,
-	list_dirs,
-	gremove,
-	gsync,
-	gsync_del,
-	remove_internal_qc_label,
-	change_gg_storage_admin_to_read_write,
-	add_verily_read_access,
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "common"))
+from release_ops import (
+    releases_df,
+    completed_platforming_raw_buckets,
+    unembargoed_dev_buckets_and_workflow_version_outputs,
+    embargoed_dev_buckets,
+)
+from gcloud_ops import (
+    list_dirs,
+    gremove,
+    gsync,
+    gsync_del,
+    remove_internal_qc_label,
+    change_gg_storage_admin_to_read_write,
+    add_verily_read_access,
 )
