+++ util/raw_bucket_prep/transfer_qc_metadata_to_raw_bucket
@@ -24,10 +24,10 @@
 import argparse
 import logging
 from pathlib import Path
-# --- resolve shared/ modules from new subfolder location ---
+
 import os, sys
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "shared"))
-from common import gsync, strip_team_prefix, gcopy
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "common"))
+from gcloud_ops import gsync, strip_team_prefix, gcopy
 from bucket_validation_utils import (
     FILE_METADATA_FILES,
     check_bucket_exists,
