+++ util/raw_bucket_prep/transfer_release_resources_to_raw_bucket.py
@@ -28,10 +28,10 @@ Authors: Javier Diaz
 import argparse
 import logging
 from pathlib import Path
-# --- resolve shared/ modules from new subfolder location ---
+
 import os, sys
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "shared"))
-from common import gcopy
+sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "common"))
+from gcloud_ops import gcopy
 import os, sys
 from collections import defaultdict
 import json
@@ -54,7 +54,7 @@ from bucket_validation_utils import (
     validate_local_release_resources_structure
 )
 
-from common import strip_team_prefix
+from gcloud_ops import strip_team_prefix
 
 logging.basicConfig(
 	level=logging.INFO,
