--- util/raw_bucket_prep/validate_raw_bucket_structure.py
@@ -1,65 +1,0 @@
-# !/usr/bin/env python3
-
-# Ensure that the raw bucket file structure is valid after contributors have
-# uploaded their data.
-# Usage: python3 validate_raw_bucket_structure.py -d team-jakobsson-pmdbs-bulk-rnaseq
-
-import argparse
-import logging
-# --- resolve shared/ modules from new subfolder location ---
-import os, sys
-sys.path.insert(0, os.path.join(os.path.dirname(os.path.abspath(__file__)), "..", "shared"))
-from common import strip_team_prefix
-from bucket_validation_utils import (
-    validate_raw_bucket_structure, 
-    check_original_metadata_files_in_bucket
-)
-
-logging.basicConfig(
-    level=logging.INFO,
-    format="%(asctime)s - %(levelname)s - %(message)s"
-)
-
-
-def main(args):
-
-    dataset_id = args.dataset_id
-    dataset_name = strip_team_prefix(dataset_id)
-    bucket_name = f"gs://asap-raw-{dataset_id}"
-    
-    logging.info(f"Validating raw bucket structure for: {bucket_name}")
-    
-    # Ensure the bucket directory structure is valid
-    validate_raw_bucket_structure(bucket_name)
-    
-    # Check metadata files in bucket. Warns if CORE tables missing, errors only if no files found.
-    files_valid = check_original_metadata_files_in_bucket(bucket_name)
-    
-    if files_valid:
-        logging.info(f"Raw bucket validation successful: {bucket_name}")
-    else:
-        logging.warning(f"Raw bucket validation completed with warnings")
-
-
-if __name__ == "__main__":
-    
-    parser = argparse.ArgumentParser(
-        description="Validate raw bucket directory structure after contributor upload"
-    )
-    
-    parser.add_argument(
-        "-d",
-        "--dataset-id",
-        type=str,
-        required=True,
-        help="Dataset ID in bucket name (e.g., team-jakobsson-pmdbs-sn-rnaseq)."
-    )
-    
-    args = parser.parse_args()
-
-    if not (args.dataset_id.startswith("team") or args.dataset_id.startswith("cohort")):
-        raise SystemExit(
-            f"--dataset-id must start with 'team' or 'cohort', got: '{args.dataset_id}'"
-        )
-
-    main(args)
