--- util/shared/common.py
@@ -1,602 +1,0 @@
-#!/usr/bin/env python3
-
-import logging
-import subprocess
-import json
-import pandas as pd
-import os
-import re
-import gspread
-from io import StringIO
-from google.cloud import storage
-from google.oauth2.service_account import Credentials
-
-
-#####################################################
-##### PULL FROM LIVE GOOGLE SPREADSHEET AS SSOT #####
-#####################################################
-SCOPES = ["https://www.googleapis.com/auth/spreadsheets.readonly"]
-
-def get_releases_df(
-	sheet_id: str = "1Qx4W3EsGQwRHXKtDd6jBnEyPGsuhxB8YCVdgJ-Mn6Hs",
-	tab_name: str = "Releases_src",
-	credentials_path: str = os.path.expanduser("~/.config/gspread/credentials.json")
-) -> pd.DataFrame:
-	if not os.path.exists(credentials_path):
-		raise FileNotFoundError(f"Credentials file not found: {credentials_path}; look at README")
-	creds = Credentials.from_service_account_file(credentials_path, scopes=SCOPES)
-	gc = gspread.authorize(creds)
-	ws = gc.open_by_key(sheet_id).worksheet(tab_name)
-	return pd.DataFrame(ws.get_all_records())
-
-releases_df = get_releases_df()
-releases_df["raw_buckets"] = "gs://asap-raw-" + releases_df["dataset_id"]
-releases_df["dev_buckets"] = "gs://asap-dev-" + releases_df["dataset_id"]
-
-ALL_TEAMS = releases_df["team_id"].unique().tolist()
-
-## Minor and Major Release that includes pipeline/curated outputs
-### The latest_workflow_version column is being used to infer datasets with pipeline outputs
-unembargoed_dev_buckets_and_workflow_version_outputs = (
-	releases_df[
-		releases_df["latest_workflow_version"].str.startswith("v", na=False)
-	]
-	.sort_values("latest_workflow_version")
-	.drop_duplicates(subset="dev_buckets", keep="last")
-	.set_index("dev_buckets")["latest_workflow_version"]
-	.to_dict()
-)
-## Urgent and Minor Release or platforming exercise during a Major Release
-completed_platforming_raw_buckets = (
-	releases_df[
-		~releases_df["latest_workflow_version"].str.startswith("v", na=False)
-	]["raw_buckets"]
-	.drop_duplicates()
-	.tolist()
-)
-
-
-######################################################################
-##### SLUG CLASSIFIERS AND ORDERED CATEGORY LISTS ####################
-######################################################################
-# Used by generate_dataset_summary_table and generate_brain_bank_summary
-# to categorize datasets by assay type, organism, and tissue source when
-# the Releases Sheet metadata isn't available (e.g., internal QC datasets).
-
-# Ordered list of assay/data-type categories. Matches the column order in
-# the dataset summary pivot tables. Add new categories here as they appear.
-ASSAY_ORDER = [
-	"sc/snRNA-seq",
-	"sc/snATAC-seq",
-	"sc/sn Multiome",
-	"Bulk RNA-seq",
-	"Spatial Transcriptomics",
-	"Proteomics",
-	"Metabolomics",
-	"Lipidomics",
-	"Genetics",
-	"WGS",
-	"Metagenomics",
-]
-
-HUMAN_SOURCES_ORDER = [
-	"Brain tissue",
-	"Cell lines",
-	"Gastrointestinal",
-	"Fecal",
-]
-
-MOUSE_SOURCES_ORDER = [
-	"Brain tissue",
-	"Liver tissue",
-	"Lung tissue",
-	"Kidney tissue",
-	"Plasma",
-	"Embryonic fibroblast",
-	"Fecal",
-]
-
-
-def team_from_slug(slug):
-	"""prod-team-hafler-pmdbs-... → hafler. Returns 'unknown' if the slug
-	does not match the expected prod-team-<name>-... shape."""
-	parts = str(slug).split("-")
-	if len(parts) >= 3 and parts[0] == "prod" and parts[1] == "team":
-		return parts[2]
-	return "unknown"
-
-
-def classify_assay(value):
-	"""Map an assay value (free-text Releases-Sheet `assay`, or a dataset slug)
-	to an entry in ASSAY_ORDER. Returns None if no pattern matches — callers
-	decide their own fallback (e.g. 'Unclassified').
-
-	The patterns cover both Releases-Sheet conventions (snake_case, free-text
-	like 'Bulk_RNA_Seq', 'mass spec') and slug conventions (kebab-case like
-	'sc-rnaseq', 'ms-mb-plasma'). Order matters: more specific patterns are
-	checked before more general ones."""
-	s = str(value).lower()
-
-	# ATAC-seq first — broader sc/sn would otherwise capture sc_atac/sn_atac
-	if any(x in s for x in [
-		"sn-atacseq", "sc-atacseq", "snatacseq", "scatacseq",
-		"sc_atac", "sn_atac", "scatac", "snatac",
-		"sc-atac", "sn-atac", "atacseq", "atac_seq", "atac-seq",
-	]):
-		return "sc/snATAC-seq"
-
-	# Multiome (paired RNA+ATAC, e.g. 10x Multiome) — checked before sc/sn RNA-seq
-	# so the multimodal/multiome keyword takes precedence over a generic sc-/sn- match
-	if any(x in s for x in ["multimodal", "multiome", "multiomic"]):
-		return "sc/sn Multiome"
-
-	# sc/sn RNA-seq — covers both kebab-case slugs and snake_case Sheet values
-	if any(x in s for x in [
-		"sn-rnaseq", "sc-rnaseq", "scrnaseq", "snrnaseq",
-		"single-cell", "single-nucleus", "single_cell", "single_nucleus",
-		"scrna", "snrna", "sc-rna", "sn-rna",
-		"sc_", "sn_",
-	]):
-		return "sc/snRNA-seq"
-
-	if "bulk" in s:
-		return "Bulk RNA-seq"
-
-	if any(x in s for x in ["spatial", "visium", "geomx", "cosmx", "xenium", "nanostring"]):
-		return "Spatial Transcriptomics"
-
-	# Mass-spec assays — tighten ms-p / ms-mb / ms-l patterns so they don't collide.
-	# Bare `ms-p` is allowed when it's the whole value (e.g. Sheet input "ms-p")
-	# or at the end of a slug; the slug-tightened `ms-p-` / `ms-p_` handles
-	# embedded positions.
-	if (any(x in s for x in ["ms-p-", "ms-p_", "proteom", "mass-spec", "mass_spec", "mass spec"])
-			or s.endswith("ms-p") or s == "ms-p"):
-		return "Proteomics"
-	if any(x in s for x in ["ms-mb-", "ms-mb_", "metabolom"]) or s.endswith("ms-mb") or s == "ms-mb":
-		return "Metabolomics"
-	if any(x in s for x in ["ms-l-", "ms-l_", "lipidom"]) or s.endswith("ms-l") or s == "ms-l":
-		return "Lipidomics"
-
-	# WGS before Genetics — Genetics is the catch-all for genotyping/SNP/array data
-	if any(x in s for x in ["wgs", "whole-genome", "whole_genome"]):
-		return "WGS"
-	if any(x in s for x in ["genetic", "genotyp", "gwas", "snp", "variant"]):
-		return "Genetics"
-	if any(x in s for x in ["metagenom", "shotgun", "microbiome", "16s"]):
-		return "Metagenomics"
-
-	return None
-
-
-def classify_organism(value):
-	"""Map an organism value (free-text Releases-Sheet `organism`, or a dataset
-	slug) to 'Human' or 'Mouse'. Returns None if neither matches.
-
-	Cell-line / in-vitro values default to Human unless the input also says
-	'mouse' (so a slug like prod-team-alessi-mefs-... resolves to Mouse via
-	the MEF check below)."""
-	s = str(value).lower()
-	# Human keywords (slug-side: pmdbs; Sheet-side: human, homo)
-	if any(x in s for x in ["human", "homo", "pmdbs"]):
-		return "Human"
-	# MEF first — must come before generic mouse check so it doesn't get
-	# swallowed by a slug that happens to contain 'mef' but not 'mouse'
-	if "mef" in s or "mefs" in s:
-		return "Mouse"
-	if any(x in s for x in ["mouse", "mus", "sulzer-fecal-metagenome-fp-spf"]):
-		return "Mouse"
-	# Cell-line / in-vitro datasets — assume Human unless the value also
-	# says mouse (caller decides further disambiguation if needed)
-	if any(x in s for x in ["invitro", "ipsc", "hek"]):
-		return "Human" if "mouse" not in s else "Mouse"
-	return None
-
-
-def classify_source(value, organism=""):
-	"""Map a tissue-source value (slug or Releases-Sheet `sample_source`) to an
-	entry in HUMAN_SOURCES_ORDER / MOUSE_SOURCES_ORDER. Returns None if no
-	pattern matches.
-
-	`organism` is optional and is used only to disambiguate Fecal samples
-	when relevant. Pass '' (the default) when classifying from a slug — Fecal
-	doesn't need disambiguation at the slug level since both Human Fecal and
-	Mouse Fecal collapse to the same 'Fecal' bucket.
-
-	When called from the Releases-Sheet path, pass the organism so Fecal can
-	be split by Human/Mouse appropriately (currently both still map to 'Fecal',
-	but the organism arg is preserved for future disambiguation needs)."""
-	s = str(value).lower()
-	o = str(organism).lower()
-
-	# MEF = mouse embryonic fibroblast → must come before generic cell-line check
-	if "mef" in s or "fibroblast" in s or "embryon" in s:
-		return "Embryonic fibroblast"
-
-	if any(x in s for x in [
-		"brain", "pmdbs", "postmortem", "midbrain", "striatum", "cortex",
-		"substantia-nigra", "substantia nigra",
-		"hippocampus", "cerebellum", "neural",
-	]):
-		return "Brain tissue"
-
-	if any(x in s for x in ["colon", "gastro", "intestin", "gi-tract", "gi tract", "gut"]):
-		return "Gastrointestinal"
-
-	# Fecal: handles both Human Fecal and Mouse Fecal. The organism arg lets
-	# callers route to bucket-specific behavior later; for now both return Fecal.
-	if any(x in s for x in ["fecal", "stool", "feces", "microbiome", "metagenom"]):
-		return "Fecal"
-
-	if "liver" in s:
-		return "Liver tissue"
-	if "lung" in s:
-		return "Lung tissue"
-	if "kidney" in s or "renal" in s:
-		return "Kidney tissue"
-	if any(x in s for x in ["plasma", "serum"]) or "-blood-" in s or s.endswith("-blood") or s == "blood":
-		return "Plasma"
-
-	if any(x in s for x in [
-		"cell line", "cell-line", "invitro", "in vitro", "ipsc", "hek",
-		"neuronal cell", "hesc", "hpsc",
-	]):
-		return "Cell lines"
-
-	return None
-
-
-######################################################################
-##### PROMOTE QC'ED METADATA AND ARTIFACTS - RAW TO PROD SECTION #####
-######################################################################
-def remove_internal_qc_label(bucket_name):
-	command = [
-		"gcloud",
-		"storage",
-		"buckets",
-		"update",
-		bucket_name,
-		"--remove-labels=internal-qc-data"
-	]
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	return result.stdout
-
-
-def get_team_name(bucket: str) -> str:
-	return bucket.split("-team-", 1)[1].split("-")[0]
-
-
-def strip_team_prefix(entity_id: str) -> str:
-    """Strip 'team' prefix if present: dataset directories do not have 'team' prefix"""
-    norm_id = entity_id.strip().lower()
-    norm_id = re.sub(r'^team[-_ ]*', '', norm_id)
-    if not norm_id:
-        raise ValueError(f"ID: [{entity_id}] is empty after stripping 'team' prefix: [{norm_id}]")
-    return norm_id
-
-
-def run_command(command):
-	try:
-		result = subprocess.run(command, check=True, capture_output=True, text=True)
-		return result.stdout
-	except subprocess.CalledProcessError as e:
-		if "No policy binding found" in e.stderr:
-			print(f"[INFO] No existing storage.admin binding to remove for {team_gg}")
-		else:
-			print(f"[ERROR] Command failed:\n{e.stderr}")
-			raise
-
-
-def check_admin_binding(bucket_name):
-	team_name = get_team_name(bucket_name)
-	role_admin = "roles/storage.admin"
-	team_gg = "asap-team-" + team_name + "@dnastack.com"
-	member = f"group:{team_gg}"
-	policy_json = run_command([
-		"gcloud",
-		"storage",
-		"buckets",
-		"get-iam-policy",
-		bucket_name,
-		"--format=json"
-	])
-	policy = json.loads(policy_json)
-	has_admin_binding = any(
-		binding["role"] == role_admin and member in binding.get("members", [])
-		for binding in policy.get("bindings", [])
-	)
-	return member, role_admin, has_admin_binding
-
-
-def change_gg_storage_admin_to_read_write(bucket_name):
-	member, role_admin, has_admin_binding = check_admin_binding(bucket_name)
-	if has_admin_binding:
-		print(f"[INFO] Removing Storage Admin access and granting Storage Object Creator and Viewer to CRN Teams for [{bucket_name}] on Google Group")
-		run_command([
-			"gcloud",
-			"storage",
-			"buckets",
-			"remove-iam-policy-binding",
-			bucket_name,
-			f"--member={member}",
-			f"--role={role_admin}"
-		])
-		run_command([
-			"gcloud",
-			"storage",
-			"buckets",
-			"add-iam-policy-binding",
-			bucket_name,
-			f"--member={member}",
-			"--role=roles/storage.objectViewer"
-		])
-		run_command([
-			"gcloud",
-			"storage",
-			"buckets",
-			"add-iam-policy-binding",
-			bucket_name,
-			f"--member={member}",
-			"--role=roles/storage.objectCreator"
-		])
-	else:
-		print(f"[INFO] Storage Object Creator and Viewer already granted to CRN Teams' permissions for [{bucket_name}] on Google Group")
-
-
-##########################################################################
-##### PROMOTE QC'ED METADATA AND ARTIFACTS - STAGING TO PROD SECTION #####
-##########################################################################
-embargoed_dev_buckets = [
-	# Human PMDBS Multimodal Seq
-	"gs://asap-raw-team-wood-pmdbs-multimodal-seq",
-	# Human PMDBS Spatial Transcriptomics Nanostring GeoMx
-	"gs://asap-dev-team-vila-pmdbs-spatial-geomx-thlc",
-	"gs://asap-dev-team-vila-pmdbs-spatial-geomx-unmasked",
-]
-
-
-def list_dirs(bucket_name):
-	command = [
-		"gcloud",
-		"storage",
-		"ls",
-		bucket_name
-	]
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	return result.stdout
-
-
-#######################################
-##### DATA INTEGRITY TEST SECTION #####
-#######################################
-def list_teams():
-	logging.info("Available teams:")
-	for team in ALL_TEAMS:
-		logging.info(team)
-
-
-def list_gs_files(bucket, release_version, workflow_name):
-	blobs = bucket.list_blobs(prefix=workflow_name) # This skips the curated metadata and artifacts directories
-	blob_names = []
-	gs_files = []
-	sample_list_loc = []
-	pattern = re.compile(rf"{workflow_name}/release/{release_version}/") # This checks for the most recent release version
-	for blob in blobs:
-		if pattern.match(blob.name):
-			blob_names.append(blob.name)
-			gs_files.append(f"gs://{bucket.name}/{blob.name}")
-			if blob.name.endswith("sample_list.tsv"):
-				sample_list_loc.append(f"gs://{bucket.name}/{blob.name}")
-	return blob_names, gs_files, sample_list_loc
-
-
-def read_manifest_files(bucket, release_version, workflow_name):
-	blobs = bucket.list_blobs(prefix=workflow_name) # This has to be called again because 'Iterator has already started'
-	manifest_dfs = []
-	pattern = re.compile(rf"{workflow_name}/release/{release_version}/")
-	for blob in blobs:
-		if blob.name.endswith("MANIFEST.tsv") and pattern.match(blob.name):
-			gs_path = f"gs://{bucket.name}/{blob.name}"
-			logging.info(f"Reading manifest: {gs_path}")
-			content = blob.download_as_text()
-			try:
-				manifest_df = pd.read_csv(StringIO(content), sep="\t")
-			except pd.errors.ParserError as e:
-				raise pd.errors.ParserError(
-					f"Failed to parse {gs_path}: {e}"
-				) from e
-			manifest_dfs.append(manifest_df)
-	combined_df = pd.concat(manifest_dfs, ignore_index=True)
-	return combined_df
-
-
-def md5_check(bucket, release_version, workflow_name):
-	blobs = bucket.list_blobs(prefix=workflow_name)
-	hashes = {}
-	pattern = re.compile(rf"{workflow_name}/release/{release_version}/")
-	for blob in blobs:
-		if pattern.match(blob.name):
-			hashes[blob] = blob.md5_hash
-	return hashes
-
-
-def non_empty_check(bucket, release_version, workflow_name, GREEN_CHECKMARK, RED_X):
-	blobs = bucket.list_blobs(prefix=workflow_name)
-	not_empty_tests = {}
-	pattern = re.compile(rf"{workflow_name}/release/{release_version}/")
-	for blob in blobs:
-		if pattern.match(blob.name):
-			if blob.size <= 10:
-				logging.error(f"Found a file less than or equal to 10 bytes: [{blob.name}]")
-				not_empty_tests[blob.name] = f"{RED_X}"
-			else:
-				not_empty_tests[blob.name] = f"{GREEN_CHECKMARK}"
-	return not_empty_tests
-
-
-def associated_metadata_check(combined_manifest_df, blob_list, GREEN_CHECKMARK, RED_X):
-	metadata_present_tests = {}
-	for file in blob_list:
-		if file.endswith("MANIFEST.tsv"):
-			metadata_present_tests[file] = "N/A"
-		else:
-			if any(file.split('/')[-1] in filename for filename in combined_manifest_df["filename"].tolist()):
-				metadata_present_tests[file] = f"{GREEN_CHECKMARK}"
-			else:
-				logging.error(f"File does not have associated metadata and is absent from MANIFEST: [{file}]")
-				metadata_present_tests[file] = f"{RED_X}"
-	return metadata_present_tests
-
-
-###########################################
-##### COMPARE STAGING TO PROD SECTION #####
-###########################################
-def compare_blob_names(results, staging):
-	staging_blob_names = results[staging]["blob_names"]
-	curated_blob_names = results["curated"]["blob_names"]
-	staging_md5_hashes = results[staging]["md5_hashes"]
-	staging_bucket_name = next(iter(staging_md5_hashes)).bucket.name
-	same_files = ["N/A"]
-	new_files = ["N/A"]
-	deleted_files = ["N/A"]
-	if sorted(staging_blob_names) == sorted(curated_blob_names):
-		logging.info(f"The blob_names in '{staging}' are equal to those in 'curated.")
-	else:
-		logging.info(f"The blob_names in '{staging}' are not equal to those in 'curated'")
-		same_files = [file for file in staging_blob_names if file in curated_blob_names]
-		new_files = [f"gs://{staging_bucket_name}/{file}" for file in staging_blob_names if file not in curated_blob_names]
-		deleted_files = [f"gs://{staging_bucket_name}/{file}" for file in curated_blob_names if file not in staging_blob_names]
-		if new_files:
-			logging.info(f"New files in '{staging}': {new_files}")
-		if deleted_files:
-			logging.info(f"Deleted files in '{staging}': {deleted_files}")
-	return same_files, new_files, deleted_files
-
-
-def compare_md5_hashes(results, staging, same_files):
-	staging_md5_hashes = results[staging]["md5_hashes"]
-	curated_md5_hashes = results["curated"]["md5_hashes"]
-	staging_bucket_name = next(iter(staging_md5_hashes)).bucket.name
-	staging_file_hashes = {key.name: value for key, value in staging_md5_hashes.items()}
-	curated_file_hashes = {key.name: value for key, value in curated_md5_hashes.items()}
-	modified_files = {}
-	for file in same_files:
-		staging_hash = staging_file_hashes.get(file)
-		curated_hash = curated_file_hashes.get(file)
-		if staging_hash and curated_hash:
-			if staging_hash != curated_hash:
-				modified_files[f"gs://{staging_bucket_name}/{file}"] = {
-					"staging_hash": staging_hash
-				}
-				logging.info(f"Modified: {file}")
-	return modified_files
-
-
-##############################################################
-##### PROMOTE WORKFLOW OUTPUTS - STAGING TO PROD SECTION #####
-##############################################################
-def gcopy(source_path, destination_path, recursive=False):
-	command = [
-		"gcloud",
-		"storage",
-		"cp",
-		source_path,
-		destination_path
-	]
-	if recursive:
-		command.insert(3, "--recursive")
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-
-	# These if's are because gcloud returns important info in stderr (e.g. "Copying /path/to/file1 to gs://bucket/file1...") 
-	# even if the command is successful. Since ERROR may be misleading, it's better to logging.info both stdout and stderr
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
-
-def gmove(source_path, destination_path):
-	command = [
-		"gcloud",
-		"storage",
-		"mv",
-		source_path,
-		destination_path
-	]
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
-
-
-def gremove(destination_path):
-	command = [
-		"gcloud",
-		"storage",
-		"rm",
-		destination_path
-	]
-	try:
-		result = subprocess.run(command, check=True, capture_output=True, text=True)
-	except subprocess.CalledProcessError:
-		logging.info(f"No files found at {destination_path}; skipping deletion.")
-		return
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
-
-
-def gsync(source_path, destination_path, dry_run):
-	command = [
-		"gcloud",
-		"storage",
-		"rsync",
-		"-r",
-		source_path,
-		destination_path
-	]
-	if dry_run:
-		command.insert(4, "--dry-run")
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
-
-
-def gsync_del(source_path, destination_path, dry_run):
-	command = [
-		"gcloud",
-		"storage",
-		"rsync",
-		"--delete-unmatched-destination-objects",
-		"-r",
-		source_path,
-		destination_path
-	]
-	if dry_run:
-		command.insert(4, "--dry-run")
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
-
-
-def add_verily_read_access(bucket_name):
-	command = [
-		"gcloud",
-		"storage",
-		"buckets",
-		"add-iam-policy-binding",
-		bucket_name,
-		"--member=group:asap-cloud-readers@verily-bvdp.com",
-		"--role=roles/storage.objectViewer",
-		"--project",
-		"dnastack-asap-parkinsons"
-	]
-	result = subprocess.run(command, check=True, capture_output=True, text=True)
-	if result.stdout:
-		logging.info(result.stdout)
-	if result.stderr:
-		logging.info(result.stderr)
